refactor: refactored the request/response headers table logic.
Previously, there is a bug in session.headers, we didn't assign the
__index metamethod to this table, this results in that all non-normalized
headers cannot be indexed.

This commit refactored this part of logic, all headers will be saved
with the lowercase and hyphen form. Also, session.headers will have the
__index metamethod (created by util.dict()).

Unfortunately there are many test cases need to be modified...